STYLE : Mypage Tabs 컴포넌트 분리
